+++ src/g_game.c
@@ -913,7 +913,7 @@ void G_ClearInput(void)
 // G_DoLoadLevel
 //
 
-static void G_DoLoadLevel(void)
+static void G_DoLoadLevel(boolean from_savegame)
 {
   int i;
 
@@ -1014,7 +1014,7 @@ static void G_DoLoadLevel(void)
 
   P_ClearDirtyArrays();
 
-  P_SetupLevel(gameepisode, gamemap, gameskill);
+  P_SetupLevel(gameepisode, gamemap, gameskill, from_savegame);
 
   MN_UpdateFreeLook();
   HU_UpdateTurnFormat();
@@ -1081,7 +1081,7 @@ static void G_ReloadLevel(void)
     G_RecordDemo(demoname_orig);
   }
 
-  G_InitNew(gameskill, gameepisode, gamemap);
+  G_InitNew(gameskill, gameepisode, gamemap, false);
   gameaction = ga_nothing;
 
   if (demorecording)
@@ -2000,7 +2000,7 @@ static void G_DoWorldDone(void)
   gamemap = wminfo.next+1;
   gamemapinfo = G_LookupMapinfo(gameepisode, gamemap);
   G_ResetRewind(false);
-  G_DoLoadLevel();
+  G_DoLoadLevel(false);
   gameaction = ga_nothing;
   viewactive = true;
   AM_clearMarks();           //jff 4/12/98 clear any marks on the automap
@@ -2276,7 +2276,7 @@ static void G_DoPlayDemo(void)
       // Do it anyway for timing demos, to reduce timing noise
       precache = timingdemo;
   
-      G_InitNew(skill, episode, map);
+      G_InitNew(skill, episode, map, false);
 
       // killough 11/98: If OPTIONS were loaded from the wad in G_InitNew(),
       // reload any demo sync-critical ones from the demo itself, to be exactly
@@ -2583,6 +2583,12 @@ static void DoSaveGame(char *name)
     // [FG] save total time for all completed levels
     JS_SetInt(doc, root_mut, "totalleveltimes", totalleveltimes);
 
+    // fast-process pending MUSINFO music change
+    if (musinfo.tics > 0)
+    {
+        musinfo.tics = 0;
+        T_MusInfo();
+    }
     // save lump name for current MUSINFO item
     char lumpname[9] = {0};
     if (musinfo.current_item > 0)
@@ -2819,7 +2825,8 @@ static boolean DoLoadGameJSON(boolean do_load_autosave, json_t *root)
     json_t *custonskilloptions_obj = JS_GetObject(root, "customskilloptions");
     LoadCustomSkillOptionsJSON(custonskilloptions_obj);
 
-    G_InitNew(gameskill, gameepisode, gamemap);
+    // do not start level music yet
+    G_InitNew(gameskill, gameepisode, gamemap, true);
 
     /* cph - MBF needs to reread the savegame options because G_InitNew
      * rereads the WAD options. The demo playback code does this too. */
@@ -2835,18 +2842,23 @@ static boolean DoLoadGameJSON(boolean do_load_autosave, json_t *root)
     totalleveltimes = JS_GetIntegerValue(root, "totalleveltimes");
 
     const char *lumpname = JS_GetStringValue(root, "musinfo");
+    int lumpnum = -1;
     if (lumpname && *lumpname)
     {
-        int lumpnum = W_CheckNumForName(lumpname);
-
-        if (lumpnum >= 0)
-        {
-            musinfo.mapthing = NULL;
-            musinfo.lastmapthing = NULL;
-            musinfo.tics = 0;
-            musinfo.current_item = lumpnum;
-            S_ChangeMusInfoMusic(lumpnum, true);
-        }
+        lumpnum = W_CheckNumForName(lumpname);
+    }
+    if (lumpnum >= 0)
+    {
+        musinfo.mapthing = NULL;
+        musinfo.lastmapthing = NULL;
+        musinfo.tics = 0;
+        musinfo.current_item = lumpnum;
+        S_ChangeMusInfoMusic(lumpnum, true);
+    }
+    else
+    {
+        // start level music now
+        S_Start();
     }
 
     max_kill_requirement = JS_GetIntegerValue(root, "max_kill_requirement");
@@ -2948,7 +2960,7 @@ static boolean DoLoadGameBinary(boolean do_load_autosave)
   LoadCustomSkillOptions(temp_p);
 
   // load a base level
-  G_InitNew(gameskill, gameepisode, gamemap);
+  G_InitNew(gameskill, gameepisode, gamemap, true); // do not start level music yet
 
   // killough 3/1/98: Read game options
   // killough 11/98: move down to here
@@ -2988,6 +3000,7 @@ static boolean DoLoadGameBinary(boolean do_load_autosave)
   }
 
   // restore MUSINFO music
+  boolean musinfo_music = false;
   if (saveg_check_size(8))
   {
       char lump[9] = {0};
@@ -3004,7 +3017,13 @@ static boolean DoLoadGameBinary(boolean do_load_autosave)
           musinfo.tics = 0;
           musinfo.current_item = lumpnum;
           S_ChangeMusInfoMusic(lumpnum, true);
+          musinfo_music = true;
       }
+  }
+  if (!musinfo_music)
+  {
+    // start level music now
+    S_Start();
   }
 
   // restore max_kill_requirement
@@ -3283,7 +3302,7 @@ void G_Ticker(void)
     switch (gameaction)
       {
       case ga_loadlevel:
-	G_DoLoadLevel();
+	G_DoLoadLevel(false);
 	break;
       case ga_newgame:
 	G_DoNewGame();
@@ -4374,7 +4393,7 @@ void G_DoNewGame (void)
   boom_basetic = gametic;             // killough 9/29/98
   true_basetic = gametic;
 
-  G_InitNew(d_skill, d_episode, d_map);
+  G_InitNew(d_skill, d_episode, d_map, false);
   gameaction = ga_nothing;
 
   if (demorecording)
@@ -4419,7 +4438,7 @@ void G_SetFastParms(int fast_pending)
 // Can be called by the startup code or the menu task,
 // consoleplayer, displayplayer, playeringame[] should be set.
 
-void G_InitNew(skill_t skill, int episode, int map)
+void G_InitNew(skill_t skill, int episode, int map, boolean from_savegame)
 {
   int i;
 
@@ -4497,7 +4516,7 @@ void G_InitNew(skill_t skill, int episode, int map)
 
   G_ResetRewind(true);
 
-  G_DoLoadLevel();
+  G_DoLoadLevel(from_savegame);
 }
 
 void G_SimplifiedInitNew(int episode, int map)
@@ -4510,7 +4529,7 @@ void G_SimplifiedInitNew(int episode, int map)
 
   G_ResetRewind(false);
 
-  G_DoLoadLevel();
+  G_DoLoadLevel(false);
 }
 
 //