+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 * **Mod Support:**
   - Add support for optional menu and intermission sounds.
+  - Add support for ambient sound definitions by SNDINFO lumps.
   - Add support for optional chaingun sound, via `DSCHGUN`.
+
 
 * **Quality of Life:**
   - Player view no longer abruptly jolts when running across very shallow floor height changes
@@ -10,6 +12,7 @@
 ## Bug Fixes
 
 * Fixed setup desktop action exiting immediately when run from the AppImage
+* Do not restart MUSINFO level music when restoring a savegame.
 
 ## Miscellaneous